+++ avr_main/src/menu/ratangular_wave.rs
@@ -2,13 +2,13 @@ use crate::microcontroler::timer::now;
 
 /// TODO: I do not know why, but to change this type to reduce ram consuption is producing a little bug
 /// which is not to blink displayed NumericalField when editing
-type X = u64;
+type X = u32;
 
 /// Generates a non-synchronous assymetric parametrizable retangular wave form
 pub struct RectangularWave<T = X> {
     up_interval: T,
     down_interval: T,
-    next_time_point: u64,
+    next_time_point: u32,
     current_state: bool,
 }
 
@@ -19,7 +19,7 @@ impl RectangularWave<X> {
         Self {
             up_interval: up_interval.clone(),
             down_interval,
-            next_time_point: now() + up_interval as u64,
+            next_time_point: now() + up_interval as u32,
             current_state: initial_state,
         }
     }
@@ -27,7 +27,7 @@ impl RectangularWave<X> {
     /// Restarts the wave form
     pub fn reset(&mut self) -> &mut Self {
         let initial_state = true;
-        self.next_time_point = now() + self.up_interval.clone() as u64;
+        self.next_time_point = now() + self.up_interval.clone() as u32;
         self.current_state = initial_state;
         self
     }
@@ -41,11 +41,11 @@ impl RectangularWave<X> {
             if self.current_state == true {
                 // up-down edge
                 self.current_state = false;
-                self.next_time_point += self.down_interval.clone() as u64;
+                self.next_time_point += self.down_interval.clone() as u32;
             } else {
                 // down-up edge
                 self.current_state = true;
-                self.next_time_point += self.up_interval.clone() as u64;
+                self.next_time_point += self.up_interval.clone() as u32;
             }
         } else {
             // no nothing
